+++ lint-staged.config.js
@@ -4,12 +4,10 @@ module.exports = {
   "*.{js,mjs,jsx}": [
     "prettier --list-different",
     "eslint --report-unused-disable-directives",
-    "git add"
   ],
   "*.{md,markdown,mdown,mkdn,mkd,mdwn,mkdown,ron}": [
     "prettier --list-different",
     "remark -f -q",
-    "git add"
   ],
-  "*.{yml,yaml}": ["prettier --list-different", "git add"]
+  "*.{yml,yaml}": ["prettier --list-different"],
 };