+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org).
 
+## 5.0.0 - 2020-05-15
+
+- Changed: minimum require `npm-package-json-lint` is now `^5.0.0`.
+- Changed: minimum require `node.js` version is `10.13.0`
+
 ## 4.0.0 - 2019-10-08
 
 - Feature: more rules.